fix(types): resolve 19 mypy regression diagnostics for PR #4 integration

+++ tests/test_filtering_processing.py
@@ -111,7 +111,7 @@ EXPECTED_CIRCULAR_I4A_HASHES = {
     ],
 )
 def test_circular_median_matches_accepted_i4a_matrix(
-    dtype: str, radius: float, expected_hash: str
+    dtype: str, radius: int, expected_hash: str
 ) -> None:
     values = np.array(
         [
